Move port_index into a separate file
